重构邮箱代码和邮件服务代码
